+++ HISTORY.rst
@@ -6,7 +6,7 @@ Release History
 0.4 (2014-08-10)
 ++++++++++++++++
 - Add Python 3 support.
-- Add inline plotting.
+- Add %inline plotting toggle.
 - Add compatibility for oct2py >= 2.0.